move print tx log processor in testscommon
